+++ src/darker/tests/test_command_line.py
@@ -1,8 +1,9 @@
-# pylint: disable=too-many-arguments
+# pylint: disable=too-many-arguments,too-many-locals
 
 """Unit tests for :mod:`darker.command_line` and :mod:`darker.__main__`"""
 
 import re
+from argparse import ArgumentError
 from importlib import reload
 from pathlib import Path
 from textwrap import dedent
@@ -288,6 +289,20 @@ def test_parse_command_line_config_src(
         expect_config=("line_length", 99),
         expect_modified=("line_length", 99),
     ),
+    dict(
+        # this is accepted as a path, but would later fail if a file or directory with
+        # that funky name doesn't exist
+        argv=["--suspicious path"],
+        expect_value=("src", ["--suspicious path"]),
+        expect_config=("src", ["--suspicious path"]),
+        expect_modified=("src", ["--suspicious path"]),
+    ),
+    dict(
+        argv=["valid/path", "another/valid/path"],
+        expect_value=("src", ["valid/path", "another/valid/path"]),
+        expect_config=("src", ["valid/path", "another/valid/path"]),
+        expect_modified=("src", ["valid/path", "another/valid/path"]),
+    ),
 )
 def test_parse_command_line(
     tmp_path, monkeypatch, argv, expect_value, expect_config, expect_modified
@@ -295,11 +310,11 @@ def test_parse_command_line(
     """``parse_command_line()`` parses options correctly"""
     monkeypatch.chdir(tmp_path)
     (tmp_path / "dummy.py").touch()
-
-    with raises_if_exception(expect_value):
+    with raises_if_exception(expect_value) as expect_exception:
 
         args, effective_cfg, modified_cfg = parse_command_line(argv)
 
+    if not expect_exception:
         arg_name, expect_arg_value = expect_value
         assert getattr(args, arg_name) == expect_arg_value
 
@@ -561,8 +576,9 @@ def test_options(git_repo, options, expect):
     dict(check=True, changes=True, expect_retval=1),
     expect_retval=0,
 )
-def test_main_retval(check, changes, expect_retval):
+def test_main_retval(git_repo, check, changes, expect_retval):
     """main() return value is correct based on --check and the need to reformat files"""
+    git_repo.add({"a.py": ""}, commit="Initial commit")
     format_edited_parts = Mock()
     format_edited_parts.return_value = (
         [
@@ -583,3 +599,32 @@ def test_main_retval(check, changes, expect_retval):
         retval = main(check_arg_maybe + ["a.py"])
 
     assert retval == expect_retval
+
+
+def test_main_missing_in_worktree(git_repo):
+    """An ``ArgumentError`` is raised if given file is not found on disk"""
+    paths = git_repo.add({"a.py": ""}, commit="Add a.py")
+    paths["a.py"].unlink()
+
+    with pytest.raises(
+        ArgumentError,
+        match=re.escape(
+            "argument PATH: Error: Path(s) 'a.py' do not exist in the working tree"
+        ),
+    ):
+
+        main(["a.py"])
+
+
+def test_main_missing_in_revision(git_repo):
+    """An ``ArgumentError`` is raised if given file didn't exist in rev2"""
+    paths = git_repo.add({"a.py": ""}, commit="Add a.py")
+    git_repo.add({"a.py": None}, commit="Delete a.py")
+    paths["a.py"].touch()
+
+    with pytest.raises(
+        ArgumentError,
+        match=re.escape("argument PATH: Error: Path(s) 'a.py' do not exist in HEAD"),
+    ):
+
+        main(["--diff", "--revision", "..HEAD", "a.py"])